+++ src/main/java/exter/foundry/tileentity/TileEntityMetalInfuser.java
@@ -113,7 +113,7 @@ public class TileEntityMetalInfuser extends TileEntityFoundryPowered {
 	}
 
 	@Override
-	public long getFoundryEnergyCapacity() {
+	public int getFoundryEnergyCapacity() {
 		return 3000;
 	}
